src/label_model/wrench_wrapper.py
"""Wrench Label Model 包装占位.
- 在真实环境下将导入 wrench 并调用其 API.
- 当前环境不可用时, 抛出 ImportError 或给予占位行为.
"""
from __future__ import annotations
from typing import Any, Dict, Optional
import numpy as np
import logging

logger = logging.getLogger(__name__)

class WrenchLabelModelWrapper:
    """Wrench标签模型包装器，处理环境依赖和错误情况."""
    
    def __init__(self, model_name: str = 'MajorityVoting', params: Optional[Dict[str, Any]] = None):
        self.model_name = model_name
        self.params = params or {}
        self._wrench_model = None
        self._available = False
        self._fitted = False
        
        # 尝试导入wrench
        try:
            import wrench
            from wrench.labelmodel import MajorityVoting, Snorkel
            self._available = True
            self._wrench = wrench
            self._model_classes = {
                'MajorityVoting': MajorityVoting,
                'Snorkel': Snorkel,
            }
            logger.info(f"Wrench可用，使用模型: {model_name}")
        except ImportError as e:
            self._available = False
            logger.warning(f"Wrench不可用: {e}，将使用占位实现")
    
    def fit(self, L, **kwargs):
        """训练标签模型."""
        if not self._available:
            logger.warning('Wrench未安装，跳过训练')
            self._fitted = True
            return
        
        try:
            # 创建模型实例
            if self.model_name in self._model_classes:
                self._wrench_model = self._model_classes[self.model_name](**self.params)
            else:
                raise ValueError(f"不支持的模型: {self.model_name}")
            
            # 训练模型
            self._wrench_model.fit(L, **kwargs)
            self._fitted = True
            logger.info(f"Wrench模型 {self.model_name} 训练完成")
            
        except Exception as e:
            logger.error(f"Wrench模型训练失败: {e}")
            self._fitted = False
            raise

    def predict_proba(self, L):
        """预测概率."""
        if not self._available:
            logger.warning('Wrench未安装，返回占位概率')
            # 占位实现：返回均匀概率
            n = L.shape[0] if hasattr(L, 'shape') else len(L)
            return np.full((n, 2), 0.5)
        
        if not self._fitted:
            raise RuntimeError("模型尚未训练，请先调用fit()")
        
        try:
            proba = self._wrench_model.predict_proba(L)
            logger.info(f"Wrench模型预测完成，形状: {proba.shape}")
            return proba
        except Exception as e:
            logger.error(f"Wrench模型预测失败: {e}")
            # 返回占位概率
            n = L.shape[0] if hasattr(L, 'shape') else len(L)
            return np.full((n, 2), 0.5)
    
    def predict(self, L):
        """预测类别标签."""
        proba = self.predict_proba(L)
        return np.argmax(proba, axis=1)
    
    @property
    def is_available(self) -> bool:
        """检查Wrench是否可用."""
        return self._available
    
    @property
    def is_fitted(self) -> bool:
        """检查模型是否已训练."""
        return self._fitted
