requirements.txt
# WASS 项目依赖
# 核心科学计算库
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 机器学习库
scikit-learn>=1.0.0

# 配置和数据处理
PyYAML>=6.0
pathlib2>=2.3.0
tabulate>=0.9.0

# 可选依赖 (弱监督学习)
# wrench-ml>=1.2.0  # 在有Wrench环境中取消注释

# 图像处理和绘图增强
pillow>=8.0.0

# 开发和测试工具 (可选)
# pytest>=6.0.0
# black>=22.0.0
# flake8>=4.0.0
